+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v.0.0.1
 
 - Paraguay service stations are added to the map.
+- Add clustering in the service stations.